--- ui/app/controllers/apps/index.js
@@ -1,12 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Controller.extend({
-  actions: {
-    destroy: function(app) {
-      if (confirm("Are you sure?")) {
-        app.deleteRecord();
-        app.save();
-      }
-    }
-  }
-});
