+++ ui/app/controllers/apps/new.js
@@ -1,13 +1,18 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
+  isFormValid: function() {
+    return Ember.isPresent(this.get("appName"));
+  }.property("appName"),
+  isFormInvalid: Ember.computed.not("isFormValid"),
+
   actions: {
     save: function() {
-      var me    = this;
-      var model = this.model;
+      var me = this;
+      var app = this.store.createRecord("app", {name: this.get("appName")});
 
-      model.save().then(function() {
-        me.transitionToRoute("apps.show.info", model);
+      app.save().then(function() {
+        me.transitionToRoute("apps.show.info", app);
       });
     }
   }
