+++ ui/app/controllers/apps/show.js
@@ -1,7 +1,8 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
-  needs: ["snow"],
+  needs: ["snow", "apps"],
   hgStatus: Ember.computed.alias("controllers.snow.model.hg_status"),
-  isHgEnabled: Ember.computed.equal("hgStatus", "enabled")
+  isHgEnabled: Ember.computed.equal("hgStatus", "enabled"),
+  apps: Ember.computed.alias("controllers.apps.model")
 });
