+++ ui/app/routes/apps/new.js
@@ -1,13 +1,7 @@
 import Ember from 'ember';
 
 export default Ember.Route.extend({
-  model: function() {
-    return this.store.createRecord('app');
-  },
-
-  deactivate: function() {
-    if (this.controller.get("model.isNew")) {
-      this.controller.get("model").rollback();
-    }
+  setupController: function(controller, model) {
+    controller.set("appName", "");
   }
 });
