--- ui/app/templates/apps.hbs
@@ -1,11 +1,0 @@
-{{#if showMenu}}
-  <ul class="nav nav-tabs">
-    {{#each app in model}}
-      {{#link-to "apps.show.info" app tagName="li" href=false current-when="apps.show"}}
-        {{link-to app.name "apps.show.info" app}}
-      {{/link-to}}
-    {{/each}}
-  </ul>
-{{/if}}
-
-{{outlet}}
