--- ui/app/templates/apps/edit.hbs
@@ -1,6 +1,0 @@
-<div class="page-header">
-  <h1>{{link-to fixedName "apps.show.info" model}}: Edit</h1>
-</div>
-
-{{partial "apps/_form"}}
-
