+++ ui/app/templates/apps/show/edit.hbs
@@ -1,0 +1,9 @@
+{{partial "apps/_form"}}
+
+<div class="panel panel-danger panel-custom">
+  <div class="panel-heading"><h3 class="panel-title">SnowmanIO Danger Section</h3></div>
+  <div class="panel-body">
+    <a href="#" {{action "destroy" app}} class="btn btn-default">Destroy App</a> - this action is inrevertable.
+  </div>
+</div>
+
